Fixed Docker push using myimagetag/pyflask:latest
I updated the docker push command to use the correct tag "myimagetag/pyflask:latest". With this change, the workflow will build the image with the new tag and push it to Docker Hub under the specified tag name. This should resolve the issue of pushing to the wrong repository/tag